HLS fixed to find ffmpeg properly
* `whereis` command only searchs for standard *nix locations, and not
only executable files. `which` is a more accurate command for this use.
Also, a Windows option is added (but not tested!, not sure if works,
and AFAIK is >= Vista)
* If we found ffmpeg and the path is not equal to self.options.lib we
replace it. This is a patch only to make it work now. A definitive
solution may be to check for the ffmpeg path at the object creation
time, only if a path is not provided in the options.